Remove old camera2 code. Fix some devices not running new ffmpeg
